add styles for login and add jsconfig
