--- .jsconfig.json
@@ -1,13 +1,0 @@
-{
-    "compilerOptions": {
-      "baseUrl": "src",
-      "module": "commonjs",
-      "target": "ES6"
-    },
-    "include": [
-      "src"
-    ],
-    "exclude": [
-      "node_modules"
-    ]
-  }
