+++ package.json
@@ -3,9 +3,10 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "bootstrap": "^5.3.3",
+    "@fortawesome/fontawesome-free": "^6.7.1",
     "cra-template": "1.2.0",
     "firebase": "^11.0.2",
+    "mdb-react-ui-kit": "^9.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^7.0.2",
