--- src/pages/Login.jsx
@@ -1,54 +1,0 @@
-import { signInWithEmailAndPassword } from 'firebase/auth';
-import React from 'react'
-import { Link, useNavigate } from 'react-router-dom';
-import { auth } from '../firebase';
-
-const Login = () => {
-
-  const [email, setEmail] = React.useState('');
-  const [password, setPassword] = React.useState('');
-
-  const navigate = useNavigate();
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-
-    try {
-      const userCredential = await signInWithEmailAndPassword(auth, email, password);
-      console.log(userCredential);
-      const user = userCredential.user;
-      localStorage.setItem('user', JSON.stringify(user));
-      localStorage.setItem('token', JSON.stringify(user.accessToken));
-      navigate('/');
-
-    } catch (error) {
-      console.log(error);
-    }
-  }
-
-  return (
-    <div>
-      <h1>Login Page</h1>
-      <form onSubmit={handleSubmit} className='login-form'>
-        <input
-          type='email'
-          placeholder='Your Email'
-          required
-          value={email}
-          onChange={(e) => setEmail(e.target.value)}
-        />
-        <input
-          type='password'
-          placeholder='Your Password'
-          required
-          value={password}
-          onChange={(e) => setPassword(e.target.value)}
-        />
-        <button type='submit' className='signup-button'>Login</button>
-      </form>
-      <p>Need to signup? <Link to='/signup'>Signup</Link></p>
-    </div>
-  )
-}
-
-export default Login
